[mppa256] Bug Fix: No GLIBC Function for / and %
I found that module and division operations using variable-saved operands
generated the use of GLIBC functions (__modsi and __udivsi3). However, when
using constants on the right side of the operators, the generated code used
MPPA-256 processor instructions. Thus, the bug was circumvented by adding a
flag that uses code with constants in the itoa and itoa64 functions.